+++ src/app/notification-badge/notification-badge-placement.demo.ts
@@ -1,0 +1,14 @@
+import {
+  Component,
+} from '@angular/core';
+
+@Component({
+  selector: 'app-notification-badge-placement-demo',
+  templateUrl: 'notification-badge-placement.demo.html',
+  styles: [`
+  .container-spacer {flex-grow: 1}
+  `]
+})
+export class NotificationBadgePlacementDemoComponent {
+
+}