+++ docs/CHANGELOG.md
@@ -10,6 +10,11 @@
 - `[Lookup]` Fixed an issue where selection for server side and paging was not working. ([#986](https://github.com/infor-design/enterprise-ng/issues/986))
 - `[Tag]` Fixed an issue where before tag remove was not called. ([#1063](https://github.com/infor-design/enterprise-ng/issues/1063))
 
+### v10.3.0 Features
+
+- `[Cards]` Added cards' selection states (single and multiple). ([#5253](https://github.com/infor-design/enterprise/issues/5253)) `EA`
+- `[Notification Badge]` Implemented the wrapper for Notification Badge Component. ([#5344](https://github.com/infor-design/enterprise/issues/5344)) `EA`
+
 ### 10.2.1 Fixes
 
 - `[Lookup]` Patch an additional issue where selection for server side and paging was not working. ([#986](https://github.com/infor-design/enterprise-ng/issues/986))